+++ packages/core/src/createActor.ts
@@ -37,17 +37,6 @@ import { toObserver } from './utils.ts';
 
 export const $$ACTOR_TYPE = 1;
 
-export type SnapshotListener<TLogic extends AnyActorLogic> = (
-  snapshot: SnapshotFrom<TLogic>
-) => void;
-
-export type EventListener<TEvent extends EventObject = EventObject> = (
-  event: TEvent
-) => void;
-
-export type Listener = () => void;
-export type ErrorListener = (error: any) => void;
-
 // those values are currently used by @xstate/react directly so it's important to keep the assigned values in sync
 export enum ProcessingStatus {
   NotStarted = 0,
@@ -833,8 +822,12 @@ export function createActor<TLogic extends AnyActorLogic>(
  * options, if any.
  *
  * @deprecated Use `createActor` instead
+ * @alias
  */
 export const interpret = createActor;
 
-/** @deprecated Use `Actor` instead. */
+/**
+ * @deprecated Use `Actor` instead.
+ * @alias
+ */
 export type Interpreter = typeof Actor;