+++ packages/core/test/types.test.ts
@@ -25,7 +25,13 @@ import {
   stateIn,
   setup,
   toPromise,
-  UnknownActorRef
+  UnknownActorRef,
+  AnyActorLogic,
+  ActorRef,
+  SnapshotFrom,
+  EmittedFrom,
+  EventFrom,
+  ActorRefFromLogic
 } from '../src/index';
 
 function noop(_x: unknown) {
@@ -398,6 +404,7 @@ it('should work with generic context', () => {
     any,
     any,
     any,
+    any,
     any, // TMeta
     any
   > {
@@ -532,6 +539,7 @@ describe('events', () => {
       _machine: StateMachine<
         TContext,
         TEvent,
+        any,
         any,
         any,
         any,
@@ -4579,4 +4587,20 @@ it('UnknownActorRef should return a Snapshot-typed value from getSnapshot()', ()
 
   // @ts-expect-error
   actor.getSnapshot().status === 'FOO';
+});
+
+it('Actor<T> should be assignable to ActorRefFromLogic<T>', () => {
+  const logic = createMachine({});
+
+  class ActorThing<T extends AnyActorLogic> {
+    actorRef: ActorRefFromLogic<T>;
+    constructor(actorLogic: T) {
+      const actor = createActor(actorLogic);
+
+      actor satisfies ActorRefFromLogic<typeof actorLogic>;
+      this.actorRef = actor;
+    }
+  }
+
+  new ActorThing(logic);
 });